Renamed hook file to match with hook's name
